+++ rules/__tests__/no-identical-title.test.js
