+++ rules/__tests__/no-skipped-tests.test.js
