+++ rules/__tests__/valid-expect-in-promise.test.js
