+++ rules/no-identical-title.js
