+++ rules/prefer-to-be-null.js
