+++ rules/valid-expect-in-promise.js
